netmap: Fix min aggregator
Do not return next element if `0` is being facing. Also, add a public key to
the test nodes. It is used in `Hash()` method and, therefore, is the main
start point in the HWR sorting, without it, every node is considered the
same, so it is hard to pick a good test value, cause sorting is random for
the same hashes.
Closes #438.

Signed-off-by: Pavel Karpy <carpawell@nspcc.ru>