netmap: Extend `NodeInfo` type with verified nodes' domain
Sometimes there is a need to control the entry of some storage nodes
into a certain group at the stage of registration in the NeoFS network.
For example, a specific organization may wish to be part of a public
network, but limit the ability of a third party to indicate its
affiliation. Previously, private node group (aka subnet) was not
implementable.

This commit extends `netmap.NodeInfo` type with getter/setter of the
verified nodes' domain name. Usage of new methods going to affect
storage nodes which try to enter the NeoFS network and use this new
option: they will have to confirm the right to do so.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>