eacl: Fix incorrect numeric rules' matching on access validation
Previously, `Validator` calculated action for eACL with numeric rules
incorrectly. This was caused by inverted comparison of filter and object
header values. For example, if rule is applied only to 'attr < 0',
the rule was applied to objects with 'attr = 1' and not applied to
objects with 'attr = -1'.

Now condition is inverted and numeric rules are matched correctly. This
also fixes corresponding unit test.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>