object/filters: Add filtering by first object ID
This field appeared for V2 split first and was not required at first glance.
More use cases have been landed now. Also, it is a successor to the SplitID in
the V1 split that does have a filter.

api-go repo is accepted for archival so no additional const importing for now,
will be fixed somehow with removing `neofs-api-go` from go.mod.

Signed-off-by: Pavel Karpy <carpawell@nspcc.ru>