netmap/policy: Replace adders with setters for list types
Еhe need to expand a partially or fully formed storage policy is either
absent or fully covered by slice manipulations. The parameter
variability is also excessive: by default the list is empty and
resetting it by calling it without parameters is difficult to read.
Adding elements one-by-one is inefficient in terms of resources.

While this is breaking change, it forces more clear app code and
performance.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>