eacl: Provide API converter checking container ID length
`NewTableFromV2` function does not check the length accepts any
container ID length:
- those that were shorter were padded with zeros;
- those that were longer were cut off.

This behavior is incorrect because it goes against the NeoFS API
protocol. However, it is not possible to fix the function backwards
compatible since it does not return an error.

Thus, a new `ReadFromV2` method is added (it is already widely used in
the library). The method checks container ID format and returns an error.
Consequently, now the eACL format is also correctly checked when
decoding a bearer token and `ContainerService.GetExtendedACLResponse`
protocol message by the API client.

`NewTableFromV2` is marked buggy and deprecated, it's going to be purged
for release later.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>