*: Affirm comparable interface of IDs
Container, object and user IDs are arrays. However, previously their
docs recommended not to treat them as arrays. In particular, this did
not allow the free use of these types with built-in `comparable`
constraint (e.g. as map keys), incl. complicating the comparison
operation, which is very popular for these entities.

Now ID types are comparable not only at the language level, but also
conceptually. Methods `Equal` are marked deprecated.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>