*: Affirm IDs as arrays
This continues cde122ecb56c893384575b51fc7997b800bc71bd and finally
commits to supporting identifiers as fixed-length byte arrays. Since
their format has not changed so far, this is unlikely to happen in the
near foreseeable future. The benefit is the user's code simplification
code and the abandonment of some no longer needed methods (they are
marked deprecated).

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>