*: Drop error from `Marshal` return
Being the interface of many types in the lib, protocol binary encoding
methods `Marshal` do not return an error because this operation cannot
go wrong. However, some methods still returned error historically. It is
worth bringing it to a general approach.

Although this change is backward incompatible, breaking the compilation
of client code is not expected: stable marshaling is mostly used for
system needs and higher level lib functions.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>