eacl: Deprecate `EqualTables` function
The function is unused. For example, `waiter.ContainerSetEACLWaiter`
compares binary representations which is correct. Overall, NeoFS
protocol does not declare comparison operation b/w two eACLs. In
particular, the eACL does not change when targets are shuffled. Thus, to
avoid introducing a synthetic property, the func will be removed soon.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>