*: Extend crypto APIs
Cryptographic signatures of digital data of the system are transmitted
in the form of a simple tuple: scheme, binary public key and the
signature itself. The library provides `Calculate` method which, on the
one hand, facilitates calculation, on the other - deprives the ability
to work with pre-calculated values. As a patch, the so-called
`StaticSigner` was provided for signing simulation.

This commit adds new approach: now the signature structure can be
flexibly created by newly added constructors (or modified by new field
setters). In some practical cases, like WalletConnect remote signing,
working with pre-computed signature structure is the only way.

For the same reason, the interface of signed data structures (e.g.
tokens) is also extended: now a structured signature setter+getter are
provided.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>