bearer: Provide way to check whether eACL is set in the token
Previously, the only way to receive eACL in the `Token` was field getter.
It states that if the table is unset, zero instance is returned. The
problem is that `eacl.Table` type does not implement `comparable`
constraint. So, there is no way to check whether eACL instance is zero
or not. Therefore, there was no way to find out whether the token
contained a table.

This adds `GetEACLTable` method returning boolean flag of eACL presence.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>