bearer: Provide way to check whether eACL is set in the token
Bearer token type provides `EACLTable` method to access underlying eACL.
It states that if the table is unset, zero instance is returned. The
problem is that `eacl.Table` type did not implement `comparable`
constraint. So, there is no way to check whether eACL instance is zero
or not. Therefore, there was no way to find out whether the token
contained a table.

This adds `eacl.Table.IsZero` method allowing to check the eACL
presence. Currently, it covers all cases where table is required incl.
bearer tokens.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>