client: Do not lose some URI parsing failure reasons on dial
Previously, `Client.Dial` lost causing error of network endpoint parsing.
Although the method returned "invalid endpoint options" error, the
original reason could not be obtained. This made debugging difficult.

Now causes are kept and returned. For this, function
`WithNetworkURIAddress` is inlined and the `ParseURI` error is not
ignored.

Fixes one of the corresponding tests.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>