internal/proto: Fix test of zero embedded message's encoding
While an unspecified (nil) embedded field of `message` protobuf type is
encoded as nothing, the zero is still reflected in the message as an
encoded field number plus zero.

This fixes the test and makes correct assertion of zero message encoding.
The tests fails meaning there is a bug for now. The test is also made
independent with the nil one to make the not fail each other.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>