proto: Fix encoding of messages with nil elements in repeated fields
Previously, implementations of `MarshaledSize` and `MarshalStable`
interfaces ignored nil elements in repeated message fields, resulting in
the following behavior: if _m_ of _n_ fields are nil, only _n-m_ ones
were encoded and further restored. Although the "nilness" of the element
makes little sense since it is not reflected in Protocol Buffers binary
format, this behavior was definitely unexpected, i.e. a bug.

This brings the solution: encode nils similarly to zeros. This is
pretty legit since the absence of a field is interpreted as zero in
protobuf. At the same time, the number of array elements will
be preserved, which is extremely important. An additional motivation was
to compare the behavior with `github.com/nspcc-dev/neofs-api-go/v2`
module, which behaves in the same way.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>