client: make Dial follow the timeout parameter
DialContext calls NewClient and it's specified to perform no I/O at all, so
we don't know whether the node is alive or not. We're querying network info
from it to ensure it's OK, but this request uses parent context, so it takes
some unknown time instead of a proper dial timeout.

Signed-off-by: Roman Khimov <roman@nspcc.ru>