client: Support signing object requests with `SignerV2`
Opens up the possibility of more flexible customization of request
signing. In the context of the new N3 scheme, this will make it easier
to switch from a (public key, signature) pairs to N3 witness scripts.

This makes client to try `SignerV2` interface in passed `Signer`
parameter. Also affects based on the client components - object slicer
and connection pool.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>