object: rework New/InitCreation
New() is absolutely useless. InitCreation() is needlessly complex. I believe
combining them gives exactly what people wanted in the first place: some way
to create object that works. For decoding new(object.Object) or
&object.Object{} are still valid and convenient, but when assembling something
for Put() to use I'd rather have this new New().

Yes, it's a breaking change, but adapting to it is trivial.

Signed-off-by: Roman Khimov <roman@nspcc.ru>